Rename config -> project_config
